requirements.txt
stardist==0.7.3
tensorflow==2.4.0
keras==2.4.3         
matplotlib==3.4.3
numba==0.54.0
Pillow==8.3.2
ipykernel==6.4.1     
scikit-image==0.18.3
opencv-python
protobuf==3.19.6    
numpy==1.20.0